backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==1.10.13
python-multipart==0.0.6
email-validator==2.1.0
typing-extensions==4.8.0

